+++ src/Core/NavGraph.cpp
@@ -50,7 +50,7 @@ void NavGraph::draw(Camera &cam_)
     if (ConfigurationManager::instance().m_debug.m_drawNavGraph)
     {
         const Vector2<float> nodeSize{5.0f, 5.0f};
-        for (int i = 0; i < m_nodes.size(); ++i)
+        for (size_t i = 0; i < m_nodes.size(); ++i)
         {
             auto &node = m_nodes[i];
             m_ren.drawRectangle(node.m_position - nodeSize / 2.0f, nodeSize, {255, 127, 39, 255}, cam_);
